+++ lib/src/domain/entity/User/data.dart
